Introduce Helper dependency

+++ entity-meta-generator-lib/src/main/java/com/loudsight/meta/serialization/transform/TemporalEntityTransform.java
@@ -44,6 +44,6 @@ public class TemporalEntityTransform extends EntityTransform<Temporal> {
         var millis = readLong(bytes);
         var nanos = readInt(bytes);
 
-        return LocalDateTime.ofEpochSecond(millis / 1000, nanos, ZoneOffset.UTC);
+        return LocalDateTime.ofEpochSecond(Double.valueOf((((double) millis))/1000.).longValue(), nanos, ZoneOffset.UTC);
     }
 }
